Multiset Generator
A multiset (or bag) is a generalization of the concept of a set that,
unlike a set, allows multiple instances of the multiset's elements.

For example, {a, a, b} and {a, b} are different multisets although
they are the same set. However, order does not matter, so {a, a, b}
and {a, b, a} are the same multiset.

The number of times an element belongs to the multiset is the
multiplicity of that member.
The total number of elements in a multiset, including repeated
memberships, is the cardinality of the multiset.

For example, in the multiset {a, a, b, b, b, c} the multiplicities of
the members a, b, and c are respectively 2, 3, and 1, and the
cardinality of the multiset is 6.

Source: https://en.wikipedia.org/wiki/Multiset

This class computes all multisets of cardinality k out of n elements.